Extract Context to doctest/parts/context.h
